+++ src/app/domain/services/lists/update-list.ts
@@ -1,0 +1,30 @@
+import { getList, updateList } from '@/db/repositories/list-repository'
+import { ListNotFoundError } from '../../errors/list-not-found-error'
+import type { InferInsertModel } from 'drizzle-orm'
+import type { schema } from '@/db/schema'
+
+export type UpdateListValues = Omit<
+  InferInsertModel<typeof schema.lists>,
+  'userId' | 'createdAt' | 'coverPath'
+>
+
+type UpdateListInput = {
+  id: string
+  userId: string
+  values: UpdateListValues
+}
+
+export async function updateListService({
+  id,
+  userId,
+  values,
+}: UpdateListInput) {
+  const [list] = await getList(id, userId)
+
+  if (!list) {
+    return new ListNotFoundError()
+  }
+
+  const [updatedList] = await updateList(id, userId, values)
+  return { list: updatedList }
+}