+++ src/db/repositories/list-repository.ts
@@ -1,8 +1,16 @@
 import { db } from '..'
-import { and, desc, eq, getTableColumns, sql } from 'drizzle-orm'
+import {
+  and,
+  desc,
+  eq,
+  getTableColumns,
+  type InferInsertModel,
+  sql,
+} from 'drizzle-orm'
 import { schema } from '../schema'
 import type { GetListsInput } from '@/app/domain/services/lists/get-lists'
 import type { InsertListModel } from '@/app/domain/entities/lists'
+import type { UpdateListValues } from '@/app/domain/services/lists/update-list'
 
 export function selectLists({
   userId,
@@ -74,4 +82,16 @@ export async function getList(id: string, userId: string) {
     .select()
     .from(schema.lists)
     .where(and(eq(schema.lists.id, id), eq(schema.lists.userId, userId)))
+}
+
+export async function updateList(
+  id: string,
+  userId: string,
+  values: UpdateListValues
+) {
+  return db
+    .update(schema.lists)
+    .set(values)
+    .where(and(eq(schema.lists.id, id), eq(schema.lists.userId, userId)))
+    .returning()
 }